starter css from Zen Garden
This is the starter css master file from Zen Garden.  I thought we could move it to a partial and look at it for a framework reference from time to time and to pull elements from.  I left it here for now to give us an idea of what projects we need to work.